+++ packages/vite-plugin-angular/src/lib/authoring/ng.spec.ts
@@ -2,12 +2,38 @@ import { compileNgFile } from './ng';
 
 const COMPONENT_CONTENT = `
 <script lang="ts">
-import { signal } from '@angular/core';
+import { signal, input, ViewChild, afterNextRender, ElementRef } from '@angular/core';
 
 defineMetadata({
-  exposes: [Math]
+  exposes: [Math],
+  queries: {
+    divElement: new ViewChild('divElement')
+  }
 });
 
+export const routeMeta = {
+  title: 'My page',
+  canActivate: [() => true],
+}
+
+export interface MyInterface {
+  title: string
+}
+
+export type MyType = string;
+
+export enum Direction {
+  Up,
+  Down,
+  Left,
+  Right,
+}
+
+export function myFunc(){
+  console.log('hello');
+}
+
+let divElement: ElementRef<HTMLDivElement>;
 let test: string;
 
 setTimeout(() => {
@@ -16,10 +42,31 @@ setTimeout(() => {
 
 const counter = signal(0);
 const [a, b, , c = 4] = [1, 2, 3];
+
+const inputWithDefault = input(""); // InputSignal<string, string>
+const inputWithoutDefault = input<string>(); // InputSignal<string | undefined, string | undefined>
+const inputWithAlias = input("", { alias: "theAlias" }); // InputSignal<string, string>
+const inputWithoutDefaultWithAlias = input<string | undefined>(undefined, {
+    alias: "theAlias",
+  }); // InputSignal<string | undefined, string | undefined>
+const inputWithTransform = input<unknown, boolean>("", {
+    transform: booleanAttribute,
+  }); // InputSignal<unknown, boolean>
+const requiredInput = input.required<string>(); // InputSignal<string, string>
+const requiredInputWithTransform = input.required<unknown, number>({
+    transform: (value) => numberAttribute(value, 10),
+  });
+const output = new EventEmitter();
+const outputWithType = new EventEmitter<string>();
+
+afterNextRender(() => {
+  console.log('the div', divElement);
+})
+
 </script>
 
 <template>
-  <div>Component</div>
+  <div #divElement>Component</div>
   <p>{{ counter() }}</p>
   <p>{ a }</p>
   <p>{ b }</p>
@@ -50,6 +97,14 @@ const elRef = inject(ElementRef);
 
 afterNextRender(() => {
   elRef.nativeElement.focus();
+});
+
+onInit(() => {
+  console.log('init code');
+});
+
+effect(() => {
+  console.log('just some effect');
 });
 </script>
 `;